pyproject.toml
[tool.poetry]
name = "naashir"
version = "0.1.0"
description = "Naashir automates the content creation workflow for translating Arabic lessons of Salafi scholars into English, streamlining text extraction, audio editing, and video creation."
authors = ["Mustafa Mohammed <mustafa.mohammed0223@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-pptx = "^1.0.2"
pywin32 = "^307"
black = "^24.8.0"
yt-dlp = "^2024.9.27"
pydantic = "^2.9.2"

[tool.poetry.scripts]
run-naashir = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
